fix: minor fixes,POSTMAN (MainService) 101/23

+++ mainApp/src/main/java/ru/practicum/ewm/user/UserServiceImpl.java
@@ -25,6 +25,9 @@ public class UserServiceImpl implements UserService {
     }
 
     public void deleteUserById(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException("User not found");
+        }
         log.info("delete user id = {}", userId);
         userRepository.deleteById(userId);
     }
